Convert Emma to two separate applications so it can be run in the dev server.
Since the dev server uses Bottle like Emma does, this change makes it possible for the
dev server to mount the Emma URLs at /_emma, so emma is available anytime the
dev server is running. In a production environment, you'll still run Emma separately
(or not at all). This change just simplifies the architecture somewhat.