Major refactoring of the caches.
I was having some very strange problems with cache files looking like their contents was duplicated (which
turned out to be because I was using class attributes instead of instance attributes - idiot!) and in the course of
investigation I discovered Python's shelve module and it seemed to be a good fit.

I could have fixed the original problem without switching to shelve, but in the end I think this design makes a
little more sense anyway. I no longer have to call into pickle myself or manage multiple cache files.