Add support for custom post metadata properties.
Previously the normalization process would ignore metadata that was
not understood, and would remove that metadata from the file during
normalization, essentially causing data loss.

Now properties that are not known will be stored in the
custom_properties dict on the Post object, and will also be
written back to the file during normalization.