Fix major bug in post breaks processor.
Previously the post breaks processor broke apart the Markdown into teaser
content and full content, then converted each resulting piece to HTML
independently. This meant that any Markdown links of the form
[link][1] would fail to be converted properly to HTML links, since the
teaser Markdown fragment didn't include the link references (they're
at the bottom of the post).

With this fix the break marker is first normalized to an HTML comment
(the Octopress-style break marker) since we know that the Markdown
processor will not touch it. Then the full post is converted to HTML
and the result is split on the break marker. The teaser content is then
stored in properly-converted HTML.