Fix Octopress-style post breaks unit test.
The unit test incorrectly checked that the content_teaser property existed, not that
it was not None.

In addition, the Octopress docs use <!--more-->, with no spaces between the comment
markers and the 'more' text. Though the regex used by the PostBreaks plugin accommodates
either, the unit test should adhere to the example used in the Octopress docs.