Possible fix for shelve sync issue
The problem is that on MacOS X calling `settings.CACHE.sync()` didn't
seem to be syncing the file. I'm not certain if there were read/writes
to the cache after this that caused problems or what, but it resulted
in a broken cache file that, at least on my machine, was exactly 96KiB.

I hacked around this to call `settings.CACHE.close()` and the cache now
seems to be written properly.

This should close tylerbutler/engineer#36

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>