Fix unittest problem on OS X
The unit tests create some temporary directories, which are then used
in some validations in the tests. In OS X, the temp directories
actually are a symlink so sometimes the paths don’t match at various
points in the code. Easily fixed by calling realpath() on the results
of mkdtemp(). More info here:

http://stackoverflow.com/questions/12482702/pythons-os-chdir-and-os-getc
wd-mismatch-when-using-tempfile-mkdtemp-on-ma