fix slug processing to allow for numeric tags
It appears that the YAML parsing is smart enough to see that something that is entirely numeric should be a number. For example:

    Tags:
    - 2013
    - 2014

However, the slugify method assumed that the `text` being passed to it was actually text. In these cases, it could be an `int` which caused `engineer build` to fail. By casting `text` to `unicode` we avoid this problem.