Ensure tags are unicode strings in all cases and add numeric tag test
Related to #80 - prior to this tags were always whatever type they were
parsed as by the YAML parser.

After this change they’ll always be unicode when accessed through the
`Post.tags` property. The `_tags` property will still contain the raw
parsed types from the PyYAML in case those are needed for some reason
in the future.

Side effect of this change: numeric tags (or any tag that would be
parsed as anything other than a unicode string) will be persisted in
the post source file as an explicit string type by the Metadata
finalization plugin. For example, a post with metadata like this:

    tags:
    - 2013

Will be rewritten as:

    tags:
    - ‘2013’