Deprecate JINJA_CACHE_DIR setting.
This setting was causing a number of unit test problems as implemented.
It makes sense to use a shared Jinja cache across all Engineer sites, unlike the other caches,
so this setting is now a constant. This is a step towards addressing #83, though it is
not a complete fix.

This change also includes the major ground work for #72 by providing a centralized
place to store Engineer app-wide settings and data.